refactor: update DB connection and user_struct.h
