+++ src/tracepoint/db.h
@@ -6,12 +6,13 @@
 #include <pqxx/pqxx>
 #include <sstream>
 #include <cstdlib>
+#include <iostream>
 #include "struct.h"
+#include "user_struct.h"
 
 class DBConnection {
 public:
     DBConnection();
-
     // event_t를 DB에 삽입하는 함수, 이벤트 목록을 batch insert 쿼리로 생성하여 삽입
     // void insert_events(const std::vector<event_t>& events);
     void insert_events(const std::vector<db_event_t>& events);
